Added support for ability of cice5 to define separate input and restart directories.

+++ payu/models/cice5.py
@@ -61,24 +61,30 @@ class Cice5(Cice):
     def set_model_pathnames(self):
         super(Cice5, self).set_model_pathnames()
 
-        # Change the INPUT path, we want everything to go into RESTART
-        self.work_input_path = self.work_restart_path
+        self.split_paths = (not self.work_input_path == self.work_restart_path)
+
+        if self.split_paths:
+            self.copy_inputs = False
 
     def archive(self):
         super(Cice5, self).archive()
 
-        res_ptr_path = os.path.join(self.restart_path, 'ice.restart_file')
-        with open(res_ptr_path) as f:
-            res_name = os.path.basename(f.read()).strip()
+        if not self.split_paths:
+            # If inputs and restarts are in the same directory have to delete old
+            # ice restarts 
 
-        assert os.path.exists(os.path.join(self.restart_path, res_name))
+            res_ptr_path = os.path.join(self.restart_path, 'ice.restart_file')
+            with open(res_ptr_path) as f:
+                res_name = os.path.basename(f.read()).strip()
 
-        # Delete the old restart file (keep the one in ice.restart_file)
-        for f in self.get_prior_restart_files():
-            if f.startswith('iced.'):
-                if f == res_name:
-                    continue
-                os.remove(os.path.join(self.restart_path, f))
+            assert os.path.exists(os.path.join(self.restart_path, res_name))
+
+            # Delete the old restart file (keep the one in ice.restart_file)
+            for f in self.get_prior_restart_files():
+                if f.startswith('iced.'):
+                    if f == res_name:
+                        continue
+                    os.remove(os.path.join(self.restart_path, f))
 
     def get_prior_restart_files(self):
         if self.prior_restart_path is not None:
